Implement platforms
This was done for the Services-Api team hackday. After spending a week
with the ISM team, I wanted to cement my learnings with the project and
the code base by implementing a new concept in ISM.

This commit contains a barebonse 'Platform' resource.

The next step which I ran out of time for was to hack a cloud
controllers /v2/service_brokers endpoint, to fetch the brokers
registered in an ism.